requirements/product.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
mysqlclient==1.4.2.post1